refactor(pilots): decompose pilot detail page into focused sub-components
- Created pilot detail components:
  - PilotOverviewTab.tsx (266L) - identity, skills, wounds, stats, awards
  - PilotCareerTab.tsx (289L) - career timeline with filters
  - PilotDetailModals.tsx (253L) - status badge, delete/edit modals
  - PilotDetailSkeleton.tsx (70L) - loading skeleton
- Reduced src/pages/gameplay/pilots/[id].tsx from 1,072L to 239L
- All 49 characterization tests passing
- Largest component: 289L (< 400L limit)
