+++ src/components/pilots/index.ts
@@ -6,3 +6,11 @@
 export { PilotCreationWizard } from './PilotCreationWizard';
 export { PilotProgressionPanel } from './PilotProgressionPanel';
 export { AbilityPurchaseModal } from './AbilityPurchaseModal';
+export {
+  StatusBadge,
+  DeleteConfirmModal,
+  EditIdentityModal,
+} from './PilotDetailModals';
+export { PilotCareerTab } from './PilotCareerTab';
+export { PilotOverviewTab } from './PilotOverviewTab';
+export { PilotDetailSkeleton } from './PilotDetailSkeleton';
